+++ src/commands/git/setup/conventional.ts
@@ -1,21 +1,21 @@
 import { Command, Flags } from '@oclif/core';
 import { Colors } from '../../../api';
 import { CLI_NAME } from '../../../shared/constants/commands';
-import { setupHusky } from '../../../modules/apps/git/setup/husky';
+import { setupConventional } from '../../../modules/apps/git/setup/conventional';
 import { globalFlags } from '@shared';
 
 export default class Release extends Command {
-  static description = 'Set up Husky for selected repositories (only for Git users)';
+  static description = 'Set up conventional commits with Husky and Commitlint for selected repositories (only for Git users)';
 
   static examples = [
-    `${Colors.PINK(`${CLI_NAME} git setup husky`)}`,
-    `${Colors.PINK(`${CLI_NAME} git setup husky -l`)}`,
+    `${Colors.PINK(`${CLI_NAME} git setup conventional`)}`,
+    `${Colors.PINK(`${CLI_NAME} git setup conventional -l`)}`,
   ];
 
   static flags = {
     ...globalFlags,
     list: Flags.boolean({
-      description: 'List all projects to select for Husky setup',
+      description: 'List all projects to select for conventional commits setup',
       char: 'l',
       default: false,
     }),
@@ -26,7 +26,7 @@ export default class Release extends Command {
       flags: { list },
     } = await this.parse(Release);
 
-    await setupHusky({
+    await setupConventional({
       list,
     });
   }