+++ src/shared/utils/code/git/index.ts
@@ -1,5 +1,5 @@
 export * from './init';
 export * from './setup/devops';
 export * from './setup/devops/templates';
-export * from './setup/husky';
+export * from './setup/conventional';
 export * from './setup';