+++ src/shared/utils/code/git/setup/index.ts
@@ -1,3 +1,3 @@
 export * from './devops';
 export * from './devops/templates';
-export * from './husky';
+export * from './conventional';