+++ src/shared/utils/code/index.ts
@@ -1,5 +1,5 @@
 export * from './git';
 export * from './git/setup/devops';
 export * from './git/setup/devops/templates';
-export * from './git/setup/husky';
+export * from './git/setup/conventional';
 export * from './git/setup';